+++ frontend/src/contexts/index.js
@@ -1,1 +1,2 @@
+export { ParticipantAddFormContext, ParticipantAddFormContextProvider } from './ParticipantAddFormContext';
 export { ParticipantContext, ParticipantContextProvider } from './ParticipantContext';
