+++ frontend/src/hooks/index.js
@@ -1,4 +1,6 @@
 export { useConfirm } from './useConfirm';
 export { useMapView } from './useMapView';
 export { useModal } from './useModal';
-export { useParticipantForm } from './useParticipantForm';
+export { useParticipantAddressSearch } from './useParticipantAddressSearch';
+export { useParticipantInputAddress } from './useParticipantInputAddress';
+export { useParticipantInputName } from './useParticipantInputName';
